--- a/src/HTTP.jl
+++ b/src/HTTP.jl
@@ -610,7 +610,7 @@ end
 include("download.jl")
 include("Servers.jl")                  ;using .Servers; using .Servers: listen
 include("Handlers.jl")                 ;using .Handlers; using .Handlers: serve
-include("parsemultipart.jl")
+include("parsemultipart.jl")           ;using .MultiPartParsing: parse_multipart_form
 include("WebSockets.jl")               ;using .WebSockets
 
 import .ConnectionPool: Transaction, Connection