--- a/test/chunking.jl
+++ b/test/chunking.jl
@@ -1,10 +1,11 @@
 using Test
-using HTTP
+using HTTP, HTTP.IOExtras
 using BufferedStreams
 
 # For more information see: https://github.com/JuliaWeb/HTTP.jl/pull/288
 @testset "Chunking" begin
     sz = 90
+    port = 8095
     hex(n) = string(n, base=16)
     encoded_data = "$(hex(sz + 9))\r\n" * "data: 1$(repeat("x", sz))\n\n" * "\r\n" *
                    "$(hex(sz + 9))\r\n" * "data: 2$(repeat("x", sz))\n\n" * "\r\n" *
@@ -15,7 +16,7 @@ using BufferedStreams
     split1 = 106
     split2 = 300
 
-    t = @async HTTP.listen("127.0.0.1", 8091) do http
+    t = @async HTTP.listen("127.0.0.1", port) do http
         startwrite(http)
         tcp = http.stream.c.io
 
@@ -34,7 +35,7 @@ using BufferedStreams
     sleep(1)
     @assert !istaskdone(t)
 
-    r = HTTP.get("http://127.0.0.1:8091")
+    r = HTTP.get("http://127.0.0.1:$port")
 
     @test String(r.body) == decoded_data
 
@@ -44,7 +45,7 @@ using BufferedStreams
         # Ignore byte-by-byte read warning
         CL = Base.CoreLogging
         CL.with_logger(CL.SimpleLogger(stderr, CL.Error)) do
-            HTTP.open("GET", "http://127.0.0.1:8091") do io
+            HTTP.open("GET", "http://127.0.0.1:$port") do io
                 io = wrap(io)
                 x = split(decoded_data, "\n")