@mbauman 2021-08-31
I can confirm this **significantly** improves the original request that was being tested/reported on Discourse — it is much larger and has an even larger benefit.

Before:

```Julia
julia> @benchmark HTTP.parse_multipart_form($r)
BenchmarkTools.Trial: 26 samples with 1 evaluation.
 Range (min … max):  181.312 ms … 235.950 ms  ┊ GC (min … max): 3.11% … 3.33%
 Time  (median):     195.296 ms               ┊ GC (median):    3.16%
 Time  (mean ± σ):   197.687 ms ±  13.029 ms  ┊ GC (mean ± σ):  3.20% ± 0.16%

 Memory estimate: 584.19 MiB, allocs estimate: 4785830.
```

After:

```Julia
julia> @benchmark HTTP.parse_multipart_form($r)
BenchmarkTools.Trial: 917 samples with 1 evaluation.
 Range (min … max):  5.131 ms …   9.564 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.309 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.447 ms ± 448.252 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

 Memory estimate: 2.03 KiB, allocs estimate: 32.
```