@quinnj 2021-08-31
Ok, I unexported `closewrite` (and friends) so nightly isn't broken due to Julia Base now exporting `closewrite`. I also fixed the websocket tests which were failing because echo.websocket.org doesnt' exist anymore. Let's see how CI fares now.